Data Export can now export more than 10000 messages In order to achieve this without overcoming Elasticsearch limits of 10000 results, more queries with 10000 row limit are aggregated. Due to this new behavior, data exported in XLS and CSV files are always sorted by Timestamp ascending, no matter the current view sort setting.

+++ console/module/data/src/main/java/org/eclipse/kapua/app/console/module/data/servlet/DataExporterExcel.java
@@ -38,7 +38,7 @@ public class DataExporterExcel extends DataExporter {
     private Workbook workbook;
     private Sheet sheet;
     private CellStyle dateStyle;
-    private short rowCount;
+    private int rowCount;
     private String topicOrDevice;
 
     private static final int MAX_ROWS = 65535;
